+++ constructor/dsl/errors.go
@@ -7,22 +7,7 @@ import (
 	"github.com/fatih/color"
 )
 
-type Error struct {
-	Line         int
-	LineContents string
-	Err          error
-}
-
-func (e *Error) Log() {
-	message := fmt.Sprintf("CONSTRUCTION FILE PARSING FAILED!\nMessage: %s\n\n", e.Err.Error())
-
-	if e.Line > 0 {
-		message = fmt.Sprintf("%sLine: %d\nLine Contents:%s\n\n", message, e.Line, e.LineContents)
-	}
-
-	color.Red(message)
-}
-
+// Parsing Errors
 var (
 	ErrCannotOpenFile = errors.New("file cannot be opened")
 	ErrEOF            = errors.New("reached end of file")
@@ -40,3 +25,22 @@ var (
 	ErrInvalidActionType              = errors.New("invalid action type")
 	ErrCannotSetVariableWithoutOutput = errors.New("cannot set variable without output path")
 )
+
+// Error contains a parsing error and context about
+// where the error occurred in the file.
+type Error struct {
+	Line         int
+	LineContents string
+	Err          error
+}
+
+// Log prints the *Error to the console in red.
+func (e *Error) Log() {
+	message := fmt.Sprintf("CONSTRUCTION FILE PARSING FAILED!\nMessage: %s\n\n", e.Err.Error())
+
+	if e.Line > 0 {
+		message = fmt.Sprintf("%sLine: %d\nLine Contents:%s\n\n", message, e.Line, e.LineContents)
+	}
+
+	color.Red(message)
+}
